changed to doubles

+++ src/motion/CoreXYKinematics.cpp
@@ -1,6 +1,6 @@
 #include "CoreXYKinematics.hpp"
 
-CoreXYKinematics::CoreXYKinematics(float steps_per_mm_) : steps_per_mm(steps_per_mm_) {}
+CoreXYKinematics::CoreXYKinematics(double steps_per_mm_) : steps_per_mm(steps_per_mm_) {}
 
 // Convert X/Y in mm to motor steps
 MotorSteps CoreXYKinematics::mm_to_steps(const XYPos& pos) const {
@@ -13,7 +13,7 @@ MotorSteps CoreXYKinematics::mm_to_steps(const XYPos& pos) const {
 // Convert motor steps to X/Y in mm
 XYPos CoreXYKinematics::steps_to_mm(const MotorSteps& steps) const {
     XYPos pos;
-    pos.x_mm = (steps.a + steps.b) / (2.0f * steps_per_mm);
-    pos.y_mm = (steps.a - steps.b) / (2.0f * steps_per_mm);
+    pos.x_mm = (steps.a + steps.b) / (2.0 * steps_per_mm);
+    pos.y_mm = (steps.a - steps.b) / (2.0 * steps_per_mm);
     return pos;
 }
