+++ src/motion/MotionSystem.hpp
@@ -10,25 +10,25 @@ using StepCallback = void(*)(bool direction);
 
 class MotionSystem {
     public:
-        MotionSystem(StepperAxis& axisA, StepperAxis& axisB, CoreXYKinematics& kinematics, float min_feature_size_mm = 1.0f);
+        MotionSystem(StepperAxis& axisA, StepperAxis& axisB, CoreXYKinematics& kinematics, double min_feature_size_mm = 1.0);
 
         // Plan and execute a linear move directly using a callback
         void moveToXY(
             const XYPos& targetPos,
-            float mm_per_s
+            double mm_per_s
         );
         void arcToXY(
             const XYPos& targetPos,
             const XYPos& centerPos,
             bool clockwise,
-            float mm_per_s
+            double mm_per_s
         );
 
     private:
         StepperAxis& _axisA;
         StepperAxis& _axisB;
         CoreXYKinematics& _kinematics;
-        float _min_feature_size_mm;
+        double _min_feature_size_mm;
 };
